% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Decide whether to change the contestants initial door selection
in the Monty Hall problem game or change their initial selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{= a true/false argument where we assign the "TRUE" value to if the contestant is sticking with their initial door selection and "FALSE" if they are staying with their initial selection.}

\item{opened.door}{= the argument which has the returned door from the open_goat_door function.}

\item{a.pick}{= the arguement containing the contestants initial door selection from the select_door function}
}
\value{
The function returns a length 1 numeric vector indictating
the contestants final door choice in the Monty Hall Problem game.
The door choice will either be doors 1, 2, or 3.
}
\description{
\code{change_door()} decides for the contestant as to whether they are
keeping their initial door selection or changing their initial
selection.
}
\details{
After the Monty Hall reveals a goat door in the "Let's Make a Deal"
game show, the contestant must make a decision as to whehter they
are going to keep their initial door selection or change their
selection to the other available door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch.
}
\examples{
  change_door( stay=T, opened.door, a.pick )

\dontrun{
change_door( stay=T, opened.door, a.pick )
}

}
