% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant won the Monty Hall Problem game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{= the argument which has the returned door from the change_door function.

@param game = the argument containing the positions of the goats and car from the create_game function}
}
\value{
The function returns a length 3 or 4 character vector
indicating whether the contestant won or loss the game.
}
\description{
\code{determine_winner()} determines if the contestants final door
selection contained a car or a goat.
}
\details{
After the contestant makes their final door selection in the
"Let's Make a Deal" game show, Monty Hall reveals whether the
contestants door contains a car or goat.
}
\examples{
  determine_winner( final.pick, game )

\dontrun{
determine_winner( final.pick, game )
}

}
