% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play an entirety of the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data framewith the results of
the Monty Hall Problem game. The data frame shows whether it
was advantegous to stay or switch in a particular Monty Hall game.
}
\description{
\code{play_game()} plays the entirety of the Monty Hall Problem
game from game creation to determining if the contestant
won or loss.
}
\details{
The game replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_game()

\dontrun{
play_game()
}

}
