% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Monty Hall Problem game a set number of times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{= The number of times the Monty Hall game will be run.}
}
\value{
The function returns a data frame with the results of
the Monty Hall game simulations. The rows contain the game strategy
information. The columns contain the game outcome information.
}
\description{
\code{play_n_games()} plays the Monty Hall Problem game a
assigned amount of times.
}
\details{
The game replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_n_games( n=100 )

\dontrun{
play_n_games( n=100 )
}

}
