% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door in a Monty Hall Problem game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector indiciting
whether the contestant selected door 1, 2, or 3.
}
\description{
\code{select_door()} selects an initial door from the three doors
presented to the contestant at the beginning of the game.
}
\details{
In the "Let's Make a Deal" game show, the contestant is presented
with three doors to choose from. Behind two doors are goats and the
other door has a car behind it.
}
\examples{
  select_door()

\dontrun{
select_door()
}


}
